+++ efinance/stock/__init__.py
@@ -7,7 +7,8 @@ from .getter import (
     get_top10_stock_holder_info,
     get_base_info,
     get_all_report_dates,
-    get_all_company_performance)
+    get_all_company_performance,
+    get_latest_holder_number)
 __all__ = ['get_history_bill',
            'get_today_bill',
            'get_latest_quote',
@@ -16,4 +17,5 @@ __all__ = ['get_history_bill',
            'get_top10_stock_holder_info',
            'get_base_info',
            'get_all_report_dates',
-           'get_all_company_performance']
+           'get_all_company_performance',
+           'get_latest_holder_number']